requirements.txt
langchain==0.3.19
loguru==0.7.2
boto3==1.36.24
fastapi[all]
fastapi-responses==0.2.1
uvicorn[standard]
gunicorn==23.0.0
openai==1.65.5
PyJWT==2.10.1
opensearch-py==2.8.0
langchain-ollama==0.2.3
requests==2.32.3
langchain-community==0.3.18
google-generativeai==0.8.4
requests-aws4auth==1.3.1
langgraph==0.3.11
langchain-openai==0.3.8
pdfplumber==0.11.6
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1